backend/requirements.txt
fastapi
langchain-openai
openai
pypdf
pdfplumber
python-docx
fpdf2
python-dotenv
python-multipart
pytest
reportlab
uvicorn[standard]
sse-starlette
filetype
slowapi
python-json-logger
diskcache
